backend/pyproject.toml
[project]
dependencies = [
"django>=5.2",
"psycopg2-binary>=2.9.9",
"djangorestframework>=3.14.0",
"gunicorn>=21.12.0",
]
name = "backend"
version = "0.1.0"
description = "A URL shortener service"
requires-python = ">=3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]

[tool.uv]
dev-dependencies = ["ruff>=0.6.2"]

[tool.hatch.build.targets.wheel]
packages = ["backend"]
